Move klee worker under klee_web
