+++ src/klee_web/api/views.py
@@ -16,8 +16,8 @@ from api.permissions import IsOwnerOrReadOnly
 from api.serializers import ProjectSerializer, FileSerializer
 from frontend.models import Project, File, Task
 
-from src.worker.worker import submit_code
-from src.worker.worker_config import WorkerConfig
+from worker.worker import submit_code
+from worker.worker_config import WorkerConfig
 
 
 class ProjectViewSet(viewsets.ModelViewSet):
