+++ src/klee_web/worker/processor/coverage.py
@@ -2,7 +2,8 @@ import os
 import glob
 
 from gcovparse import gcovparse
-from src.worker.processor.base import BaseProcessor
+
+from worker.processor.base import BaseProcessor
 
 
 class CoverageProcessor(BaseProcessor):
