+++ src/klee_web/worker/processor/failed_test.py
@@ -1,7 +1,8 @@
 import linecache
 import os
 import re
-from src.worker.processor.base import BaseProcessor
+
+from worker.processor.base import BaseProcessor
 
 
 class FailedTestProcessor(BaseProcessor):
