+++ src/klee_web/worker/processor/klee_run.py
@@ -1,4 +1,4 @@
-from src.worker.processor.base import BaseProcessor
+from worker.processor.base import BaseProcessor
 
 
 class KleeRunProcessor(BaseProcessor):
